Add outline to main(), import code from raytrace
Add an outline of what the code should be able to do to main() to assist in
code development.  When writing (code, papers, novels, etc.), a clear outline
helps greatly with the creative process.

Imported the functional code from raytrace into rays.c and mirrors.c, along
with the header information placed into sd_defs.h.  Work will need to be done
to generalize this code to make it work within the context of scopedesign
(rather than the specific application for which raytrace was written).

	modified:   Makefile.am
	modified:   main.c
	modified:   mirrors.c
	new file:   mirrors.h
	modified:   rays.c
	modified:   rays.h
	modified:   sd_defs.h