First pass at using DS9 by main()
Convert check for DS9 from presence check to actually returning the path to
DS9.  Add this information to the config.h.in file for use in the actual
application.

Add this information to main() to allow the main program to call ds9, and play
with placing this call into a separate thread.

	modified:   config.h.in
	modified:   configure.ac
	modified:   src/main.c