Add ARGTABLE2 library to the SCOPEDESIGN repository.
Since it is less likely to be installed on a user's system, and is more of an
internal functionality enchancement for SCOPEDESIGN, add ARGTABLE (v2.13) to
the SCOPEDESIGN repository.  This library is used for parsing command-line
arguments, and is small enough to tag along with the remainder of the source
code.

In the current form, the ARGTABLE2 shared library and headers are installed on
the system during the "make install" phase, but this comes after the "make all"
phase whereby the executables are created and linked to the library.  This will
require some tweaking.

	modified:   Makefile.am
	modified:   config.h.in
	modified:   configure.ac
	new file:   libargtable/AUTHORS
	new file:   libargtable/CMakeLists.txt
	new file:   libargtable/COPYING
	new file:   libargtable/ChangeLog
	new file:   libargtable/Makefile.am
	new file:   libargtable/Makefile.in
	new file:   libargtable/Makefile.nmake
	new file:   libargtable/README
	new file:   libargtable/arg_date.c
	new file:   libargtable/arg_dbl.c
	new file:   libargtable/arg_end.c
	new file:   libargtable/arg_file.c
	new file:   libargtable/arg_int.c
	new file:   libargtable/arg_lit.c
	new file:   libargtable/arg_rem.c
	new file:   libargtable/arg_rex.c
	new file:   libargtable/arg_str.c
	new file:   libargtable/argtable2.c
	new file:   libargtable/argtable2.def
	new file:   libargtable/argtable2.h
	new file:   libargtable/config.h.in
	new file:   libargtable/getopt.c
	new file:   libargtable/getopt.h
	new file:   libargtable/getopt1.c