Set up ARGTABLE to work as a "convenience library".
Since ARGTABLE2 is unlikely to be installed on a user's system, and since it is
small and not subject to changes and updates as frequently as, say, GSL or
CFITSIO, set up ARGTABLE2 to be used as a "convenience library".  There is no
longer any check for its presence in configure.ac, and the src/Makefile.am has
been alerted to the the presence and requirement of libargtable2.a within the
repository.

The EXTRA_DIST files in libargtable/ need to be updated to reflect the library's
use as a "convenience library" for this project.

	modified:   config.h.in
	modified:   configure.ac
	modified:   libargtable/Makefile.am
	modified:   src/Makefile.am
	modified:   src/main.c