Spiffy up the rays_reflect business and other main() cleanup
Cleanup of main() to place ARGTABLE stuff into separate function at the end
of the main.c file.  Add testing code to main() related to reflections.

The order of solving for the components of o in rays_reflect() depends on the
direction of the normal vector to the mirror surface.  Add three subfunctions
for the three cases of primary n-hat at that point on the optic.  Make the main
rays_reflect() function have a status return value that can be checked.

	modified:   src/main.c
	modified:   src/rays.c
	modified:   src/rays.h