Lots of work... display with DS9, ray initialization, write to FITS
Now able to talk with DS9 to display images and affect how DS9 shows data.  This
uses the XPA library, distributed by SAO.  It may be beneficial to add XPA as
a "convenience library" in the way that ARGTABLE is in the code base now, so
that users don't have to install it separately.

Ability to write ray positions to FITS is now live.  At present, the initial
starting positions of the rays are written to a test FITS file, which could
be displayed using DS9.

Ray initialization has also been partially implimented, with skeleton code for
other ideas.

Functionality increases!  This will be the final form of v 0.1.1.  The next
commit will bump the in-progress version to 0.2.0.

	modified:   config.h.in
	modified:   configure.ac
	modified:   src/display.c
	modified:   src/display.h
	modified:   src/illumination.c
	modified:   src/illumination.h
	modified:   src/main.c
	modified:   src/rays.c
	modified:   src/rays.h
	modified:   src/sd_defs.h