Added functionality to open and control a single DS9 window
Added new code to display.c to open and control a single DS9 window.  Code
cleanup will be necessary, still, but functionality is present!  The crux of
the issue is to extract the port information for the desired DS9 window and
pass that information to XPA so that only the given window is affected.

Collateral changes in main.c and sd_defs.h for this added functionality.

Version bump to 0.2.0 (change in configure.ac).

	modified:   configure.ac
	modified:   src/display.c
	modified:   src/display.h
	modified:   src/main.c
	modified:   src/sd_defs.h