Streamlining the DS9/XPA experience
With the working DS9 window operation, work to streamline the proecess, and to add cases
for how to open the window.  For instance, do we force a new window, canibalize an
existing window, or do whatever?

Next steps will be to add XPA as a local "convenience library" so that a separate
installation of XPA is no longer necessary.

	modified:   config.h.in
	modified:   configure.ac
	modified:   src/display.c
	modified:   src/display.h
	modified:   src/main.c