XPA now included as a "convenience library"
The XPA library (version 2.1.17) is now included in the ScopeDesign repository
as a "convenience library", so end-users do not have to have XPA installed
separately on their systems.  Only the parts of XPA v.2.1.17 needed for the
"shared library" part of the application have been included here -- there is
additional functionality included in the full XPA release (available at
https://github.com/ericmandel/xpa).

To-Do: add/modify a readme file in the libxpa/ directory describing the version
of XPA included and the caveats above about only the files necessary for the
"shared library" version are included.

	modified:   Makefile.am
	modified:   config.h.in
	modified:   configure.ac
	new file:   libxpa/BUGS
	new file:   libxpa/COPYRIGHT
	new file:   libxpa/INSTALL
	new file:   libxpa/LICENSE
	new file:   libxpa/Makefile.am
	new file:   libxpa/README
	new file:   libxpa/acl.c
	new file:   libxpa/client.c
	new file:   libxpa/clipboard.c
	new file:   libxpa/command.c
	new file:   libxpa/find.c
	new file:   libxpa/find.h
	new file:   libxpa/gtkloop.c
	new file:   libxpa/port.c
	new file:   libxpa/remote.c
	new file:   libxpa/tcl.c
	new file:   libxpa/tclloop.c
	new file:   libxpa/tcp.c
	new file:   libxpa/tcp.h
	new file:   libxpa/timedconn.c
	new file:   libxpa/timedconn.h
	new file:   libxpa/word.c
	new file:   libxpa/word.h
	new file:   libxpa/xalloc.c
	new file:   libxpa/xalloc.h
	new file:   libxpa/xlaunch.c
	new file:   libxpa/xlaunch.h
	new file:   libxpa/xpa.c
	new file:   libxpa/xpa.h
	new file:   libxpa/xpaio.c
	new file:   libxpa/xpap.h
	new file:   libxpa/xport.h
	new file:   libxpa/xtloop.c
	modified:   src/Makefile.am