Completely functional opening of DS9 window
Code now opens or canibalizes a DS9 window in the function display_open_ds9()
without error.  Uses the three symbolic integers DS9_FORCE_NEW, DS9_CANIBALIZE,
and DS9_WHATEVER properly, with the latter two checking for open windows first
and taking appropriate action.

This function sets the external variable ds9_port with the port # required by
the XPA routines for later communication.  This variable is available to all
routines in display.c, which is the only set of functions that need to call
XPA.

	modified:   src/display.c