Add SPLASH screen / text to terminal
Add a splash text output to the terminal when the program opens.  The function
display_splash() is written to be extensible with displaying an actual splash
screen in the future with GTK (once I get around to learning how).  For now,
the splash text outputs the application version number, and the verion numbers
of the important external libraries used.

	modified:   src/display.c
	modified:   src/display.h
	modified:   src/main.c