Add the program XPANS to the distribution, needed to register DS9 ports
Add the program XPANS from the XPA distribution to the list for build and
install.  This program is required for DS9 to be able to register its ports
and be accesses using the libxpa.a routines.  Add the program file and update
the Makefile.am to properly build and install it.

	modified:   libxpa/Makefile.am
	new file:   libxpa/xpans.c