Tweaks to display_ds9_open() to ensure we control only the desired DS9 window
When opening a new window, use a unique title for that DS9 window tagged with
the TTY slot information (provided by the libc routine ttyslot()), so that
if the application is being used by multiple users on a single machine under a
single account (not unlikely with Camps), any given instance of the application
will control only the DS9 window that _it_ creates.  Savvy?

	modified:   src/display.c