Initialize geometry of the telescope
Added new files demo.(hc) containing the parameters for demonstration
telescopes.  So far, they are populated with the specs for a Newtonian
reflector.  The idea is that these demo routines will provide a basis for the
development of the ScopeDesign code, and will remain in the code base for
demonstration purposes, but that eventually telescope geometry will be input
using ASCII *.txt files or command line input or even GTK+ widgets.  The
framework for using alternative inputs is available in the function
setup_initialize_geometry(), which is called directly from main().

Some shakeout will be required of this scheme as scode development progresses,
but it is the hope that these initial steps will prove to be foundational.

Next, code that initializes the illumination environment will be required.

	modified:   src/Makefile.am
	new file:   src/demo.c
	new file:   src/demo.h
	modified:   src/main.c
	modified:   src/sd_defs.h
	modified:   src/setup.c
	modified:   src/setup.h