Build system cleanup, version bump, added XPA standalones
Version bump to 0.2.2.  Not really sure about the version scheme at this point,
since code is in such gritty development.  No specific plan or direction for
any given version, but with no one actually paying attention, does it really
matter?  Likely not.

Clean up the configure.ac file a bit to clarify which header and function checks
are required for each part of the project.  The goal, perhaps, will be to use
getopt() functionality instead of LIBARGTABLE for argument processing -- so
clarify which things are associated with LIBARGTABLE for axing at the
appropriate time.

Add XPAGET and XPASET standalone programs to the LIBXPA part of the
installation, since these are useful command-line tools, particularly calling
on all DS9 windows to quit.

	modified:   config.h.in
	modified:   configure.ac
	modified:   libxpa/Makefile.am
	new file:   libxpa/xpaget.c
	new file:   libxpa/xpaset.c
	modified:   src/sd_defs.h