Allow N_RAYS to be dynamically configurable depending on memory
Changed N_RAYS from being a static definition in sd_defs.h to being a global
variable that may be dynaically configurable using routines in the new init.c
source file.  The idea is that N_RAYS should be adjustable based on the amount
of RAM in the system and amount free.

Ultimately, the code will loop over the ray-tracing seciton several times to
build up signal-to-noise in the final output images, and the number of loops
can be adjustable based on N_RAYS.

	modified:   ../config.h.in
	modified:   ../configure.ac
	modified:   Makefile.am
	modified:   demo.c
	modified:   display.c
	modified:   illumination.c
	new file:   init.c
	new file:   init.h
	modified:   main.c
	modified:   mirrors.c
	modified:   rays.c
	modified:   sd_defs.h
	modified:   setup.c
	new file:   ui.c
	modified:   vectors.c