Add RAM-checking system call (for Mac, linux coming), GTK+ window
Add system call to check for amount of RAM on the system to use as a guide for
how many rays to initialize at one time.  If more rays are required, then the
code will make more and add them to the appropriate FITS images along the way.
This code is in place for the Mac... linux coming.

Also, added basic GTK+ window, with the idea being that said window can be the
interactive point for the program to adjust telescope type, illumination type,
etc.  It will also have a "COMPUTE" button to actually run the code,
eventually.

Need to figure out differences between GTK+ 3 and 2 and what to do if a system
has neither.

	modified:   config.h.in
	modified:   configure.ac
	modified:   src/Makefile.am
	modified:   src/display.c
	modified:   src/init.c
	modified:   src/init.h
	modified:   src/main.c
	modified:   src/sd_defs.h
	modified:   src/ui.c
	modified:   src/ui.h