Add Linux memory check; set N_RAYS based on available RAM
Add check for available memory in Linux (in addition to extant Mac check).  Use
that value to compute N_RAYS to fill that space.  Criteria is to use no more
than 4 GB of RAM for a cycle of rays, or 1/4 of the system memory, whichever is
smaller.

The idea will be to allow the code to loop through the ray-tracing multiple
times in order to build up the proper S/N in the various downstream optical
planes.

	modified:   src/init.c