Remove dependency on TPEB-Library, version bump
Version 0.2.4.  Removed dependency on the tpeb-library, since we were only
using a couple routines.  Furthermore, when CFITSIO is upgraded on a machine,
it requires the tpeb-library to be recompiled first, then scopedesign... a bit
of a pain.

Added the two source files from tpeb-library that are used presently, along
with the associated headers.  Modified the illumination.c source file to point
to the new source files rather than the tpeb-library.  Added the new source
files to the Makefile.am, and removed the TPEB-LIBRARY checks from the
configure.ac file.

Whew!

	modified:   config.h.in
	modified:   configure.ac
	modified:   src/Makefile.am
	modified:   src/display.c
	new file:   src/fitsw.c
	new file:   src/fitsw.h
	modified:   src/illumination.c
	new file:   src/imutil.c
	new file:   src/imutil.h