Add missing XPA file needed to compile library.
When compiling on a Raspberry Pi, (with no previous XPA installation), found
that the XPA file prsetup.h was required but missing from the libxpa/
directory included with the ScopeDesign repository.

Removed the XPA-default INSTALL file, as it is not relevant to the context
here as a "convenience library".  Updated the README to include information
about the version of XPA included with ScopeDesign and what the XPA-default
README was below.

	deleted:    INSTALL
	modified:   README
	new file:   prsetup.h