New branch for complete re-write of CFITSIO wrapper functions
Existing CFITSIO wrapper functions are from pre-existing TPEB library
(circa 2009-10).  This branch is intended for a complete re-write of the
wrapper functions based on a more thorough understanding of FITS files,
and to create wrapper functions specifically tailored for the needs of
ScopeDesign.

This branch is meant to be independent of the functional development of
ScopeDesign in terms of the ray tracing and input/output work, and will
be merged into the master (and active development branches) upon
satisfactory completetion of the new functions.

New functions will be written into the fitswrap.c and fitswrap.h files.

	new file:   fitswrap.c
	new file:   fitswrap.h